Extract objectMapper bean into its own config file to allow re-use in test which was broken because of objectMapper not being included in test context
