+++ jparest/src/main/java/uk/gov/homeoffice/digital/sas/jparest/Config.java
@@ -1,14 +1,14 @@
 package uk.gov.homeoffice.digital.sas.jparest;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.datatype.hibernate5.Hibernate5Module;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import org.springdoc.core.customizers.OpenApiCustomiser;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.transaction.PlatformTransactionManager;
+import uk.gov.homeoffice.digital.sas.jparest.config.ObjectMapperConfig;
 import uk.gov.homeoffice.digital.sas.jparest.exceptions.exceptionhandling.ApiResponseExceptionHandler;
 import uk.gov.homeoffice.digital.sas.jparest.swagger.PathItemCreator;
 import uk.gov.homeoffice.digital.sas.jparest.swagger.ResourceOpenApiCustomiser;
@@ -17,17 +17,10 @@ import javax.persistence.EntityManager;
 import java.util.logging.Logger;
 
 @Configuration
+@Import({ObjectMapperConfig.class})
 public class Config {
 
     private static final Logger LOGGER = Logger.getLogger(Config.class.getName());
-
-    @Bean
-    public ObjectMapper objectMapper() {
-        ObjectMapper objectMapper = new ObjectMapper();
-        objectMapper.registerModule(new Hibernate5Module());
-        objectMapper.registerModule(new JavaTimeModule());
-        return objectMapper;
-    }
 
     @Bean
     @Lazy(false)
